# docs/openapi.yaml
openapi: "3.0.0"
info:
  title: AME Wallet API
  description: |
    A API AME Wallet representa uma carteira virtual, com ela você gerenciará seus recursos financeiros relizar azer compras, transaferências entre outras operações.
  version: "2.0.0"
paths:
  /info:
    get:
      summary: Retorna informações da carteira
      description: Retornar as informações relacionadas a carteira principal do owner. A carteira e o owner são identificados apartir do header ```Authorization```.
      operationId: getInfo
      security:
      - OAuth2: [read]
      tags:
      - Gerenciamento de Carteiras
      responses:
        '200':
          description: Informações encontradas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: Informações não encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Falha na autenticação. Autenticação negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Falha na autorização. Você não possui permissão para realizar esta operação.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 'Método HTTP ou Content-Type não suportado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro interno no servidor"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - Gerenciamento de Carteiras
      operationId: putInfo
      security:
      - OAuth2: [write]
      summary: Atualiza informações da carteira
      description: Após uma carteira ser criada é possível alterar os dados cadastradas, desde de que as mudanças respeitam as mesmas regras da criação.
      parameters:
      - in: path
        name: walletId
        required: true
        schema:
          type: string

      requestBody:
        description: "Modelo de requisição utilizado para criar/atualizar uma carteira"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRequest'
      responses:
        '200':
          description: Carteira atualizada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Falha na autenticação. Autenticação negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Falha na autorização. Você não possui permissão para realizar esta operação.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 'Método HTTP ou Content-Type não suportado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro interno no servidor"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /wallets:
    description: "A `wallet` é o principal objeto da API, com ela você será capaz de gerenciar todos os seus recursos financeiros assim como realizar diversar operações, tais como: compras, transferéncia, recarrargas receber cashbacks."
    post:
      tags:
      - Gerenciamento de Carteiras
      operationId: postWallets
      security:
      - OAuth2: [write]
      summary: Cria Carteira
      description: Cria uma nova carteira associada a um `onwer`. O `Owner` representa o dono/responsável pela carteira. Através da carteira virtual será possível utilizar todos os recursos da AME Wallet API. O Owner será criado juntamente com a sua respectiva carteira. Caso o owner já exista será lançada uma exceção, o sistema identifica owner apartir do número de documento cadastrado logo não podem existir dois iguais.
      requestBody:
        description: "Modelo de requisição utilizado para criar uma nova carteira"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRequest'
      responses:
        '201':
          description: Carteira criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Falha na autenticação. Autenticação negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Falha na autorização. Você não possui permissão para realizar esta operação.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 'Método HTTP ou Content-Type não suportado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro interno no servidor"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  '/orders/{orderId}/refunds/{refundId}':
    put:
      tags:
      - Ordens
      security:
      - OAuth2: [write]
      summary: Realiza o estorno de uma ordem.
      description: Realiza o estorno referente a uma ordem já realizada.
      parameters:
      - in: path
        name: orderId
        schema:
          type: string
        required: true
        description: Id da ordem que deseja realizar o estorno.
      - in: path
        name: refundId
        schema:
          type: string
        required: true
        description: ID do estorno.
      requestBody:
        description: "Modelo de requisição de uma ordem de estorno"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundOrderRequest'
      responses:
        '200':
          description: "Ordens retornadas com sucesso."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  '/orders/{orderId}/release':
    post:
      tags:
      - Ordens
      operationId: postOrderRelease
      security:
      - OAuth2: [write]
      summary: Pré Autoriza uma liquidação de uma ordem
      description: Realiza a liquidacão de uma ordem.
      parameters:
      - in: path
        name: orderId
        schema:
          type: string
        required: true
        description: Id da ordem que deseja realizar a aliquidação
      requestBody:
        description: "Modelo de quisição de uma ordem de liquidação"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseOrderRequest'
      responses:
        '201':
          description: Liquidação executada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseOrderResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      - error_description
      properties:
        error:
          description: Classificação do erro
          type: string
          example: wallet_validation
        error_description:
          description: Descrião do erro
          type: string
          example: Já existe uma carteira cadastrada a um owner com este documento.
        error_id:
          description: Identificador do erro
          type: string
          example: "WALLET_0001"
    BalanceResponse:
      type: object
      required:
      - available
      - futureCredit
      - futureDebit
      properties:
        available:
          type: integer
          format: int64
          description: "Saldo disponível para consumo. O valor é representado em centavos, por exemplo 59539 é equivalente á R$ 595,39"
          example: 59539
        futureCredit:
          type: integer
          format: int64
          description: "Saldo que será creditado em uma data futura. O valor é representado em centavos, por exemplo 59539 é equivalente á R$ 595,39"
          example: 59539
        futureDebit:
          type: integer
          format: int64
          description: "Saldo que será debitado em uma data futura. O valor é representado em centavos, por exemplo 59539 é equivalente á R$ 595,39"
          example: 59539
    OwnerRequest:
      type: object
      required:
      - name
      - email
      - document
      - documentType
      - externalId
      properties:
        name:
          type: string
          description: Nome do `owner`
          example: Frederico da Silva
        externalId:
          type: string
          description: Id do owner no sistema de origem externo. Utilizado para armazenar um id externo no caso de uma itegração.
          example: "1HYgb3443X"
        email:
          type: string
          description: Email do `onwer` no momento do registro.
          example: 'frederico.silva@amedigical.com'
        document:
          type: string
          description: O valor do documento cadastrado no momento do registro do `onwer`
          example: '999-999-999-99'
        documentType:
          type: string
          description: É a constante representa qual o tipo de documento que está associado ao `onwer`.
          enum:
          - CPF
          - CNPJ
    WalletRequest:
      type: object
      required:
      - name
      - type
      - owner
      properties:
        name:
          type: string
          description: Nome da wallet
          example: Carteira Principal
        type:
          type: string
          description: Tipo da carteira, pode possuir os seguintes valores ```CUSTOMER``` ou ```MERCHANT```. Os tipos de carteira definem um conjunto de operação e limites associados.
          enum:
          - CUSTOMER
          - MERCHANT
          example: CUSTOMER
        main:
          type: boolean
          description: Indica se a carteira criada será a carteira principal do owner. Caso owner ainda não exista obrigatóriamente a cartiera será considerada como principal. Um owner não pode possuir duas carteiras principais.
          default: true
          example: true
        owner:
          $ref: '#/components/schemas/OwnerRequest'
    OwnerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do `owner`.
          example: 'f030bca2-bff3-4531-a8c0-a33f676022d6'
          readOnly: true
        externalId:
          type: string
          description: Id do owner no sistema de origem externo. Utilizado para armazenar um id externo no caso de uma itegração.
          example: "1HYgb3443X"
        name:
          type: string
          description: Nome do `owner`
          example: Frederico da Silva
        email:
          type: string
          description: Email do `onwer` no momento do registro.
          example: 'frederico.silva@amedigical.com'
        document:
          type: string
          description: O valor do documento cadastrado no momento do registro do `onwer`
          example: '999-999-999-99'
        documentType:
          type: string
          description: É a constante representa qual o tipo de documento que está associado ao `onwer`.
          enum:
          - CPF
          - CNPJ
        createdAt:
          type: string
          format: date-time
          description: Data e hora da criação da owner no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-17T13:22:41.SSS -03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da última atualização da owner no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-27T13:22:41.SSS -03:00'
    WalletResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '632c4f16-51b7-46a0-9377-89dcfc307bf1'
          description: Identificador único da `wallet`
          readOnly: true
        type:
          type: string
          enum:
          - CUSTOMER
          - MERCHANT
          example: CUSTOMER
        name:
          type: string
          description: Carteira de Investimentos.
        main:
          type: boolean
          description: Entre as varias carteiras que um `owner` pode possuir, indica qual é a principal.
          example: true
        active:
          type: boolean
          description: Inidica se a carteira está ativa para utilização.
        createdAt:
          type: string
          format: date-time
          description: Data e hora da criação da carteira no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-17T13:22:41.SSS -03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da última atualização da carteira no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-27T13:22:41.SSS -03:00'
        owner:
          $ref: '#/components/schemas/OwnerResponse'
        balance:
          $ref: '#/components/schemas/BalanceResponse'
...
    RefundOrderRequest:
      allOf:
      - $ref: '#/components/schemas/OrderRequest'
      - type: object
        properties:
          type:
            type: string
            example: REFUND
            enum: [REFUND]
    ReleaseOrderRequest:
      type: object
      properties:
        type:
          type: string
          example: RELEASE
          default: RELEASE
        takeRate:
          type: number
          format: float
          example: 1.5
          description: Representa a taxa cobrada pela liquidação da ordem.
          default: 0.0
        releaseTime:
          type: integer
          example: 30
          description: Quantidade de um determinado perioro até a data de liquidação.
        releaseTimeUnit:
          type: string
          enum:
          - MONTH
          - DAY
          - YEAR
          default: DAY
          description: Determina qual a unidade de tempo utilizada para cálcular a data de liquidação junto com o `releaseTime`.
    ReleaseOrderResponse:
      type: object
      properties:
        id:
          type: string
          description: Identificador único da ordem de liquidação
        type:
          type: string
          example: RELEASE
          default: RELEASE
        status:
          type: string
          enum:
          - AUTHORIZED
          - RELEASED
        takeRate:
          type: number
          format: float
          example: 1.5
          description: Representa a taxa cobrada pela liquidação da ordem.
          default: 0.0
        releaseTime:
          type: integer
          example: 30
          description: Quantidade de um determinado perioro até a data de liquidação.
        releaseTimeUnit:
          type: string
          enum:
          - MONTH
          - DAY
          - YEAR
          default: DAY
          description: Determina qual a unidade de tempo utilizada para cálcular a data de liquidação junto com o `releaseTime`.
        action:
          type: string
          description: "Representa a ação realizada que resultou no status atual da transação"
          enum:
          - AUTHORIZE
          - RELEASE
        createdAt:
          type: string
          format: date-time
          description: Data e hora da criação da ordem no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-17T13:22:41.SSS -03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da última da ordem no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-27T13:22:41.SSS -03:00'
        transactions:
          type: array
          items:
            oneOf:
            - $ref: '#/components/schemas/CreditcardTransactionResponse'
            - $ref: '#/components/schemas/AmeCashTransactionResponse'
            discriminator:
              propertyName: type
...
    AmeCashTransactionResponse:
      allOf:
      - $ref: '#/components/schemas/TransactionResponse'
      - type: object
        required:
        - paymentMethod
        properties:
          paymentMethod:
            type: string
            enum: [AME_CASH]
    OrderResponse:
      type: object
      required:
      - id
      - type
      - status
      - title
      - amountInCents
      - transactions
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "bb84ee76-95da-4ebe-a944-263a2a0343dc"
          description: Identificador único da ordem
        type:
          type: string
          example: PURCHASE
          enum:
          - PURCHASE
          - CASH_IN
          - TRANSFER_BETWEEN_WALLETS
        status:
          type: string
          enum:
          - CREATED
          - AUTHORIZED
          - CAPTURED
          - CANCELLED
        action:
          type: string
          description: "Representa a ação realizada que resultou no status atual da ordem"
          enum:
          - CREATE
          - AUTHORIZE
          - CAPTURE
          - CANCEL
          - SYS_CANCEL
        title:
          type: string
          example: 'Deposito com cartão de crédito'
          description: Título da ordem
        description:
          type: string
          example: 'Deposito realizado para comprar utilizando saldo da carteira.'
          description: 'Descrição da ordem'
        amountInCents:
          type: integer
          format: int64
          example: 157990
          minimum: 1
          description: Valor total da ordem em centavos, onde o `amount` deve ser maior que zero.
        transactionType:
          type: string
        paymentMethods:
          type: array
          description: "Meios de pagamento que foram utilizados para executar a ordem."
          items:
            type: string
            enum:
            - CREDITCARD
            - AME_CASH
        transactions:
          type: array
          items:
            oneOf:
            - $ref: '#/components/schemas/CreditcardTransactionResponse'
            - $ref: '#/components/schemas/AmeCashTransactionResponse'
            discriminator:
              propertyName: type

        createdAt:
          type: string
          format: date-time
          description: Data e hora da criação da ordem no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-17T13:22:41.SSS -03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da última da ordem no padrão `yyyy-MM-ddTHH:mm:ss Z`
          example: '2018-08-27T13:22:41.SSS -03:00'
...
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        password:
          tokenUrl: https://api.dev.amedigital.com/auth/v1/platforms/ame/oauth/token
          scopes:
            read: Concede acesso de leitura
            write: Concede acesso de escrita
        clientCredentials:
          tokenUrl: https://api.dev.amedigital.com/auth/v1/platforms/ame/oauth/token
          scopes:
            read: Concede acesso de leitura
            write: Concede acesso de escrita